+++ packages/react-storage/src/components/StorageBrowser/displayText/libraries/en/shared.ts
@@ -25,6 +25,7 @@ export const DEFAULT_ACTION_VIEW_DISPLAY_TEXT: Omit<
   tableColumnNameHeader: 'Name',
   tableColumnTypeHeader: 'Type',
   tableColumnSizeHeader: 'Size',
+  tableColumnProgressHeader: 'Progress',
 };
 
 export const DEFAULT_LIST_VIEW_DISPLAY_TEXT: Omit<