+++ packages/react-storage/src/components/StorageBrowser/views/LocationActionView/DeleteView/DeleteView.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
+import { classNames } from '@aws-amplify/ui';
 
 import { ViewElement } from '../../../components/elements';
 import { ActionCancelControl } from '../../../controls/ActionCancelControl';
+import { ActionConfirmationModalControl } from '../../../controls/ActionConfirmationModalControl';
 import { ActionExitControl } from '../../../controls/ActionExitControl';
 import { ActionStartControl } from '../../../controls/ActionStartControl';
 import { DataTableControl } from '../../../controls/DataTableControl';
@@ -14,7 +16,6 @@ import { STORAGE_BROWSER_BLOCK } from '../../../components';
 import { DeleteViewProvider } from './DeleteViewProvider';
 import { useDeleteView } from './useDeleteView';
 import type { DeleteViewType } from './types';
-import { classNames } from '@aws-amplify/ui';
 
 export const DeleteView: DeleteViewType = ({ className, ...props }) => {
   const state = useDeleteView(props);
@@ -39,6 +40,7 @@ export const DeleteView: DeleteViewType = ({ className, ...props }) => {
             <ActionStartControl />
           </ViewElement>
         </ViewElement>
+        <ActionConfirmationModalControl />
       </DeleteViewProvider>
     </ViewElement>
   );
@@ -49,6 +51,7 @@ DeleteView.displayName = 'DeleteView';
 DeleteView.Provider = DeleteViewProvider;
 
 DeleteView.Cancel = ActionCancelControl;
+DeleteView.ConfirmationModal = ActionConfirmationModalControl;
 DeleteView.Exit = ActionExitControl;
 DeleteView.Message = MessageControl;
 DeleteView.Start = ActionStartControl;