+++ packages/react-storage/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/ui-react-storage
 
+## 3.16.0
+
+### Minor Changes
+
+- [#6834](https://github.com/aws-amplify/amplify-ui/pull/6834) [`0401829108ffae82db4fa776c96e8d0075a591b9`](https://github.com/aws-amplify/amplify-ui/commit/0401829108ffae82db4fa776c96e8d0075a591b9) Thanks [@osama-rizk](https://github.com/osama-rizk)! - feat(storage-browser): Folder deletion.
+
+### Patch Changes
+
+- [#6845](https://github.com/aws-amplify/amplify-ui/pull/6845) [`a003cb5507d973834e83f5bd908ebcfb78eff1b8`](https://github.com/aws-amplify/amplify-ui/commit/a003cb5507d973834e83f5bd908ebcfb78eff1b8) Thanks [@osama-rizk](https://github.com/osama-rizk)! - fix(react-storage): handle empty string prefix in controlled mode
+
+- Updated dependencies [[`0401829108ffae82db4fa776c96e8d0075a591b9`](https://github.com/aws-amplify/amplify-ui/commit/0401829108ffae82db4fa776c96e8d0075a591b9)]:
+  - @aws-amplify/ui-react-core@3.6.0
+  - @aws-amplify/ui-react@6.15.0
+  - @aws-amplify/ui@6.15.0
+
 ## 3.15.0
 
 ### Minor Changes